Move BufferSupplier to utils package to avoid cyclic dependencies
