+++ clients/src/test/java/org/apache/kafka/common/compress/KafkaLZ4Test.java
@@ -18,9 +18,7 @@ package org.apache.kafka.common.compress;
 
 import net.jpountz.xxhash.XXHashFactory;
 
-import org.apache.kafka.common.compress.KafkaLZ4BlockInputStream;
-import org.apache.kafka.common.compress.KafkaLZ4BlockOutputStream;
-import org.apache.kafka.common.record.BufferSupplier;
+import org.apache.kafka.common.utils.BufferSupplier;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
