+++ clients/src/test/java/org/apache/kafka/common/record/BufferSupplierTest.java
@@ -17,6 +17,7 @@
 
 package org.apache.kafka.common.record;
 
+import org.apache.kafka.common.utils.BufferSupplier;
 import org.junit.jupiter.api.Test;
 
 import java.nio.ByteBuffer;
