+++ raft/src/main/java/org/apache/kafka/raft/internals/RecordsBatchReader.java
@@ -18,7 +18,7 @@ package org.apache.kafka.raft.internals;
 
 import org.apache.kafka.common.protocol.DataInputStreamReadable;
 import org.apache.kafka.common.protocol.Readable;
-import org.apache.kafka.common.record.BufferSupplier;
+import org.apache.kafka.common.utils.BufferSupplier;
 import org.apache.kafka.common.record.DefaultRecordBatch;
 import org.apache.kafka.common.record.FileRecords;
 import org.apache.kafka.common.record.MemoryRecords;
